New file structure + some code refactoring
